Use async enumerable for cosmos db repository
